requirements.txt
selenium
webdriver-manager
requests
websocket-client
